Add deprecated core.game_state shim with migration logger
